pyproject.toml
[project]
name = "researchops-studio"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0,<3",
  "pydantic-settings>=2.0,<3",
  "PyYAML>=6.0",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
  "langgraph>=0.2,<0.3",
  "langchain-core>=0.3.58,<1.0",
  "PyJWT[crypto]>=2.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "httpx>=0.25",
  "ruff>=0.4",
  "black>=24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
